[Decomp][Unit] Cohesion split: extract damage dealing/scaling to UnitDamage.cpp
